Introduce new rc_support module
A test can call `init_runtime_config` during it's initialization
and the code will look for a `.advancedtestsrc` file in the current
working directory. If the file is found and there is a section matching
the specified `namespace` argument all values will be added
(non-overriding) to `os.environ` preserving case.